college_event_management/backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.1
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
Pillow==10.2.0
django-allauth==0.63.3
requests-oauthlib==1.3.1
dj-rest-auth==6.0.0
python-decouple==3.8
django-filter==24.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0
